requirements.txt
numpy
scikit-learn
tensorflow-cpu
boto3
aws-lambda-powertools
